# src/main/resources/schema/swagger/example-api-swagger-spec.yml
swagger: '2.0'
info:
  version: 0.0.1
  title:  Example Api
  description: ''
host: 'localhost:8080'
basePath: /example-api
tags:
  - name: Purchases
    description: 'Purchase a product'
paths:
  '/v1/purchase':
    post:
      operationId: purchaseProduct
      summary: 'Purchase a product'
      tags:
        - Purchases
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/purchase-save-request'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/response-base'
        '401':
          $ref: '#/responses/response-unauthorized'
        '500':
          $ref: '#/responses/response-internal-server-error-500'
  '/v1/purchases/{customer-id}':
      get:
        operationId: getAllTransactions
        summary: Get All transactions for a customer
        tags:
          - Purchases
        produces:
        - application/json
        parameters:
          - $ref: '#/parameters/customer-id-in-path'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-model-payment-transaction-list'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
  '/v1/purchases/{transaction-id}/refund':
      put:
        operationId: refundPayment
        summary: 'Refund payment'
        tags:
          - Purchases
        consumes:
        - application/json
        parameters:
          - $ref: '#/parameters/transaction-id-in-path'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-base'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
parameters:
  transaction-id-in-path:
    in: path
    name: transaction-id
    type: string
    required: true
  customer-id-in-path:
    in: path
    name: customer-id
    type: string
    required: true
definitions:
    response-base:
      type: object
      properties:
        success:
          type: boolean
        resultCode:
          type: integer
          format: int32
        message:
          type: string
        errors:
          type: array
          items:
            type: string
        transactionId:
          type: string
        type:
          type: string
      discriminator: type
      required:
        - success
        - resultCode
    purchase-save-request:
      type: object
      properties:
        phoneNumber:
          type: string
        productId:
          type: string
        amount:
          type: number
          format: double
      required: 
        - phoneNumber
        - productId
        - amount
    model-payment-transaction:
      type: object
      properties:
        phoneNumber:
          type: string
        productId:
          type: string
        transactionId:
          type: string
        amount:
          type: number
          format: double
        transactionBegin:
          type: string
          format: date-time
        transactionEnd:
          type: string
          format: date-time
    model-payment-transaction-list:
      type: object
      properties:
        list:
          type: array
          items: 
            $ref: '#/definitions/model-payment-transaction'
    response-model-payment-transaction-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-payment-transaction-list'
responses:
    response-internal-server-error-500:
      description: Internal Server error
      schema:
        $ref: '#/definitions/response-base'
    response-unauthorized:
        description: Authentication information is missing or invalid
        schema:
          $ref: '#/definitions/response-base' 
        headers:
          WWW_Authenticate:
            type: string